More bug fixes

+++ Infrastructure/Services/Ytdlp/YtdlpJsonClasses/Video.cs
@@ -4,6 +4,6 @@ public class Video
 {
     public string? Id { get; set; }
     public string? Title { get; set; }
-    public int? Duration { get; set; }
+    public float? Duration { get; set; }
     public string? Uploader { get; set; }
 }
